rename: cardInfoContext 이름 변경
